[BOTF-196] Add logic functions to helper files
## [BOTF-196] Description
Created the helper files
