+++ src/components/inputs/listDocPicker/ListDocPicker.scss
@@ -14,6 +14,7 @@
          cursor: pointer;
          -webkit-user-select: none;
          user-select: none;
+         margin-left: 0;
       }
 
       .doc-list-item:hover {
