+++ src/style/scss/typography.scss
@@ -74,7 +74,7 @@ ul {
    }
 }
 
-li:not(.MuiListItem-root) {
+li {
    margin-left: 1rem;
 }
